>> Почистил main.c
